fix: [alert] allow decimal for alert threshold value (#17751)
* fix: [alert] add tooltip message for alert threshold value

* support decimal value for alert condition threshold

* add integration test

+++ superset/reports/schemas.py
@@ -103,7 +103,7 @@ class ValidatorConfigJSONSchema(Schema):
         description=validator_config_json_op_description,
         validate=validate.OneOf(choices=["<", "<=", ">", ">=", "==", "!="]),
     )
-    threshold = fields.Integer()
+    threshold = fields.Float()
 
 
 class ReportRecipientConfigJSONSchema(Schema):
